Create `DatastoreStorageFactory` using `spine-gcloud-testutil`.
